basic db and no session
