+++ package.json
@@ -24,9 +24,11 @@
     "typescript": "^5.2.2"
   },
   "dependencies": {
+    "connect-mongo": "^5.1.0",
     "crc": "^4.3.2",
     "express": "^4.18.2",
     "express-winston": "^4.2.0",
+    "mongoose": "^8.6.0",
     "server": "file:",
     "winston": "^3.10.0"
   },
