+++ src/controllers/ticketShopController.ts
@@ -14,5 +14,5 @@ export const list = (req: Request, res: Response) => {
       }
     ]
   };
-  encryptAndSend(data, res);
+  encryptAndSend(data, res,req);
 };
