+++ src/sample.config.ts
@@ -1,3 +1,5 @@
 //TODO replace this file with environment variables .env file and sample.env
 export const IP = "127.0.0.1";
 export const PORT = 80;
+export const DB_ADDRESS = "mongodb://root:example@${IP}:27017";
+export const DB_NAME = "bormoir"
