+++ src/types/index.ts
@@ -1,0 +1,8 @@
+declare namespace Express {
+    export interface Request {
+        gameSession?: {
+            session_id: string;
+            uu_id?: string;
+          }[];
+    }
+ }
